requirements.txt
flask
langchain-pinecone
langchain-cohere
python-dotenv
cohere
nltk
pinecone
pypdf
